+++ CHANGELOG.md
@@ -6,8 +6,16 @@ All notable changes to this project will be documented in this file.
 - `0.2.x` Releases - [0.2.0] [0.2.1]
 - `0.3.x` Releases - [0.3.0]
 - `0.4.x` Releases - [0.4.0] [0.4.1] [0.4.3] [0.4.4]
+- `0.5.x` Releases - [0.5.0]
 
 ---
+
+## [0.5.0](https://github.com/space-bunny/node-sdk/releases/tag/v0.5.0)
+
+#### Changed
+
+- Improve support for endpoint configs
+- Emit events for connections (connect, disconnect, close, error)
 
 ## [0.4.4](https://github.com/space-bunny/node-sdk/releases/tag/v0.4.4)